feat: update filter to moodle 4.5
