+++ subfilters/client.php
@@ -24,7 +24,7 @@
  * @copyright  WIRIS Europe (Maths for more S.L)
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
-class filter_wiris_client extends moodle_text_filter {
+class filter_wiris_client extends \core_filters\text_filter {
